+++ Common/Gfx/Util/D2DUtil.cpp
@@ -33,7 +33,7 @@ D2D1_POINT_2F AddPoint2F(const D2D1_POINT_2F& point1, const D2D1_POINT_2F& point
 	return D2D1::Point2F(point1.x + point2.x, point1.y + point2.y);
 }
 
-D2D1_POINT_2F FindEdgePoint(const UINT32 theta, const float left, const float top, const float width, const float height)
+D2D1_POINT_2F FindEdgePoint(const float theta, const float left, const float top, const float width, const float height)
 {
 	double theta1 = theta * (M_PI / 180.0f);
